feat: add products images

+++ database/migrations/2023_10_26_164938_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('slug')->unique();
 
             $table->text('content')->nullable();
-            $table->decimal('price')->nullable();
+            $table->integer('price')->nullable();
 
             $table->string('image')->nullable();
